+++ .gitignore
@@ -14,3 +14,4 @@ results
 
 node_modules
 npm-debug.log
+netlify-cms-github-oauth-provider.zip
