+++ lambda.js
@@ -1,0 +1,9 @@
+const app = require('./app')
+
+const awsServerlessExpress = require('aws-serverless-express')
+const awsServerlessExpressMiddleware = require('aws-serverless-express/middleware')
+
+app.use(awsServerlessExpressMiddleware.eventContext())
+
+const server = awsServerlessExpress.createServer(app)
+exports.handler = (event, context) => awsServerlessExpress.proxy(server, event, context)
